first step into a new refactor, simplying calcs and removing duplicates
